// src/store/types.ts
export type NodeType = 'corner' | 'smooth' | 'offcurve' | 'qcurve'
// ...
export interface SelectedSegmentState {
  pathId: string
  startNodeId: string
  endNodeId: string
  type: 'line' | 'quad' | 'cubic' | 'quadBlob'
}